feat: add catalog route and sidebar component
