+++ src/components/Sidebar/sidebar.scss
@@ -1,0 +1,1 @@
+@import '../../constants.scss';
