+++ src/routes/Catalog/catalog.scss
@@ -1,0 +1,1 @@
+@import '../../constants.scss';
